Some more atari driver shuffling. Added atarigen_init() function which allocates all timers and registers for save states. Updated all drivers to call it. Converted some drivers to use device timers.

+++ src/mame/includes/xybots.h
@@ -11,7 +11,7 @@ struct _xybots_state
 {
 	atarigen_state	atarigen;
 
-	int 			h256;
+	UINT16 			h256;
 };
